+++ src/components/index.js
@@ -7,10 +7,10 @@ import TeamSection from "./about-us/TeamSection";
 import EventCard from "./cards/EventCard";
 import EventCardMobile from "./cards/EventCardMobile";
 import EventModal from "./cards/EventModal";
-import ResourcesList from "./resources/ResourcesList";
-import ResourcesFeaturedDesktop from "./resources/ResourcesFeaturedDesktop";
-import ResourcesFeaturedMobile from "./resources/ResourcesFeaturedMobile";
-import ResourcesFeatured from "./resources/ResourcesFeatured";
+import ResourcesList from "./resources/resourcesList/ResourcesList";
+import ResourcesFeaturedDesktop from "./resources/featured/ResourcesFeaturedDesktop";
+import ResourcesFeaturedMobile from "./resources/featured/ResourcesFeaturedMobile";
+import ResourcesFeatured from "./resources/featured/ResourcesFeatured";
 import ResourcesCard from "./cards/ResourcesCard";
 import AddCalendar from "./events/AddCalendar";
 import CustomButton from "./buttons/CustomButton";
