+++ src/components/resources/featured/ResourcesFeatured.js
