+++ src/components/resources/form/ResourceFormDesktop.js
@@ -2,10 +2,10 @@ import React from "react";
 import { Formik, Form, Field } from "formik";
 import * as Yup from "yup";
 
-import FormikField from "../FormikField/FormikField"
+import FormikField from "../../FormikField/FormikField"
 // import "../components/form.css"
 import { CheckboxWithLabel } from 'formik-material-ui';
-import FileUploadBtn from '../FormikField/FileUploadBtn'
+import FileUploadBtn from '../../FormikField/FileUploadBtn'
 import Button from '@material-ui/core/Button';
 
 import Grid from '@material-ui/core/Grid';
@@ -13,10 +13,10 @@ import Grid from '@material-ui/core/Grid';
 
 import classNames from "classnames";
 import { makeStyles } from "@material-ui/core/styles";
-import styles from "../../assets/material-kit-assets/jss/material-kit-react/views/landingPage.js";
-import {CustomHeader, Template} from ".."
+import styles from "../../../assets/material-kit-assets/jss/material-kit-react/views/landingPage.js";
+import { CustomHeader } from "../.."
 import Container from '@material-ui/core/Container';
-import firebase from "../../firebase";
+import firebase from "../../../firebase";
 
 import Categories from "./FormCategories"
 const MainCategories = Categories.FormCategories;
@@ -232,37 +232,46 @@ const ResourceFormDesktop = (props) => {
     }
 
     return (
-        <Template title={'Add New Resource'} active={"resources"} brand={"VIRTUAL CAMPUS"}>
         <div>
+            <CustomHeader active={'resources'} brand={"VIRTUAL CAMPUS"}></CustomHeader>
             <div className={classNames(classes.mainOther, manual.main)}>
-                <Container>
-
-
-                    <div style={{
-                        fontFamily: "Poppins", fontStyle: "normal", fontWeight: "normal",
-                        fontSize: "36px", lineHeight: "54px", color: "#0072CE"
-                    }}>
-                            <div style={{
-                                fontFamily: "Poppins", fontStyle: "normal", fontWeight: "normal",
-                                fontSize: "36px", lineHeight: "54px", color: "#0072CE"
-                            }}>
-                                Add a New Resource
-                            </div>
-                            <div style={{
-                                fontFamily: "Poppins", fontStyle: "normal", fontWeight: "normal",
-                                fontSize: "14px", lineHeight: "21px"
-                            }}>
+                <Container maxWidth='lg' style={{ paddingTop: '76px' }}>
+                    {/* <Grid container spacing={8}>
+                        <Grid item xs={4}>
+                            <div className={classNames(manual.toAll, manual.title)}>Add a New Resource</div>
+                            <div className={classNames(manual.toAll, manual.detail)}>
                                 Thank you for your interest in sharing your project through CVC.
                                 Please fill out the following form so we can thoroughly promote your resource on our website!
                             </div>
-                            <div style={{
-                                fontFamily: "Poppins", fontStyle: "normal", fontWeight: "normal",
-                                fontSize: "14px", lineHeight: "21px", paddingTop: "66px"
-                            }}>
-                                Questions? Contact us at <br />
-                                <a href='mailto:columbiavirtualcampus@gmail.com'>columbiavirtualcampus@gmail.com</a>.
+                            <div className={classNames(manual.toAll, manual.detail)} style={{ paddingTop: '87px' }}>
+                                Questions? Contact us at: <a style={{ textAlign: "center", color: "#4284C8" }}
+                                    href={"mailto:columbiavirtualcampus@gmail.com"}>columbiavirtualcampus@gmail.com</a>.
                             </div>
-                    </div>
+                        </Grid> */}
+                        <Grid container spacing={10}>
+                            <Grid item xs={4}>
+                                <div style={{
+                                    fontFamily: "Poppins", fontStyle: "normal", fontWeight: "normal",
+                                    fontSize: "36px", lineHeight: "54px", color: "#0072CE"
+                                }}>
+                                    Add a New Resource
+                          </div>
+                                <div style={{
+                                    fontFamily: "Poppins", fontStyle: "normal", fontWeight: "normal",
+                                    fontSize: "14px", lineHeight: "21px"
+                                }}>
+                                    Thank you for your interest in sharing your project through CVC.
+                                    Please fill out the following form so we can thoroughly promote your resource on our website!
+                          </div>
+                                <div style={{
+                                    fontFamily: "Poppins", fontStyle: "normal", fontWeight: "normal",
+                                    fontSize: "14px", lineHeight: "21px", paddingTop: "66px"
+                                }}>
+                                    Questions? Contact us at <br />
+                                    <a href='mailto:columbiavirtualcampus@gmail.com'>columbiavirtualcampus@gmail.com</a>.
+                          </div>
+                        </Grid>
+                        <Grid item xs={8}>
                             <Formik initialValues={initVal} onSubmit={submitHandler} validationSchema={validationSchema}>
                                 {({ dirty, isValid, errors, touched }) => {
                                     return (
@@ -270,10 +279,10 @@ const ResourceFormDesktop = (props) => {
                                             <div className={manual.section}>
                                                 <div className={classNames(manual.toAll, manual.subtitle)}>Contact</div>
                                                 <Grid container spacing={2}>
-                                                    <Grid item xs={12}>
+                                                    <Grid item sm={6}>
                                                         <FormikField label="Name / Organization" name="name" error={errors.name} touch={touched.name} required ></FormikField>
                                                     </Grid>
-                                                    <Grid item xs={12}>
+                                                    <Grid item sm={6}>
                                                         <div>
                                                             <FormikField label="Email" name="email" error={errors.email} touch={touched.email} required ></FormikField>
                                                         </div>
@@ -284,53 +293,50 @@ const ResourceFormDesktop = (props) => {
                                             <div className={manual.section}>
                                                 <div className={classNames(manual.toAll, manual.subtitle)} style={{ marginTop: '30px' }}>Resource</div>
                                                 <Grid container spacing={2} >
-                                                    <Grid item xs={12}>
+                                                    <Grid item sm={6}>
                                                         <FormikField label="Project Name" name="project_name" error={errors.project_name} touch={touched.project_name} required ></FormikField>
                                                     </Grid>
-                                                </Grid>
-                                                <Grid container spacing={2} >
-                                                    <Grid item xs={12}>
+                                                    <Grid item sm={6}>
                                                         <div>
                                                             <FormikField
                                                                 label="Logo / Image Link"
                                                                 name="image_link"
                                                                 error={errors.image_link}
                                                                 touch={touched.image_link}
-                                                                // value={this.state.imgFileValue}
+                                                            // value={this.state.imgFileValue}
                                                             />
                                                         </div>
                                                     </Grid>
-                                                </Grid>
                                                     {/*<Grid item sm={2}>*/}
-                                                        {/* <Field component={SimpleFileUpload} name="file" className="input-image" label="Image Upload" /> */}
-                                                        {/* <CustomButton text={"Upload File"} color={"blue"} size={"small"}/> */}
-                                                        {/* <Field
-                                                            name='image_file'
-                                                            component={FileUpload}
-                                                        /> */}
-                                                        {/* <input
-                                                            name='image_file'
-                                                            type='file'
-                                                            id='file_upload'
-                                                            style={{ display: 'none' }}
-                                                        />
-                                                        <label htmlFor="file_upload">
-                                                            <div>
-                                                                <Button className={classNames(manual.toAll, manual.uploadBtn)} variant="outlined" component="span">
-                                                                    Upload File
-                                                                </Button>
-                                                            </div>
-                                                        </label> */}
-                                                    {/*<FileUploadBtn
-                                                            text="Upload"
-                                                            name='file'
-                                                            label='Image Upload'
-                                                            id="fileUpload"
-                                                            // onChange={this.imgFileUploadHandler}
-                                                        />
+                                                    {/*    /!* <Field component={SimpleFileUpload} name="file" className="input-image" label="Image Upload" /> *!/*/}
+                                                    {/*    /!* <CustomButton text={"Upload File"} color={"blue"} size={"small"}/> *!/*/}
+                                                    {/*    /!* <Field*/}
+                                                    {/*        name='image_file'*/}
+                                                    {/*        component={FileUpload}*/}
+                                                    {/*    /> *!/*/}
+                                                    {/*    /!* <input*/}
+                                                    {/*        name='image_file'*/}
+                                                    {/*        type='file'*/}
+                                                    {/*        id='file_upload'*/}
+                                                    {/*        style={{ display: 'none' }}*/}
+                                                    {/*    />*/}
+                                                    {/*    <label htmlFor="file_upload">*/}
+                                                    {/*        <div>*/}
+                                                    {/*            <Button className={classNames(manual.toAll, manual.uploadBtn)} variant="outlined" component="span">*/}
+                                                    {/*                Upload File*/}
+                                                    {/*            </Button>*/}
+                                                    {/*        </div>*/}
+                                                    {/*    </label> *!/*/}
+                                                    {/*    <FileUploadBtn*/}
+                                                    {/*        text="Upload"*/}
+                                                    {/*        name='file'*/}
+                                                    {/*        label='Image Upload'*/}
+                                                    {/*        id="fileUpload"*/}
+                                                    {/*    // onChange={this.imgFileUploadHandler}*/}
+                                                    {/*    />*/}
 
-                                                    </Grid>*/}
-                                                    {/* </Grid> */}
+                                                    {/*</Grid>*/}
+                                                </Grid>
 
                                                 <Grid container spacing={2}>
                                                     <Grid item xs={12}>
@@ -348,14 +354,14 @@ const ResourceFormDesktop = (props) => {
                                                         <div className={manual.dot} style={{ paddingTop: '9px' }}>•</div>
                                                         <div style={{ marginLeft: '35px', paddingTop: '9px', height: '15px', width: '70px' }}>Category</div>
                                                     </Grid>
-                                                    <Grid item sm={2}>
+                                                    <Grid item sm={10}>
                                                         <div className="buttons">
                                                             {/* {Object.keys(MainCategories).map(category => {
                                                                 return (
                                                                     <Button name="category" className={classNames(manual.toAll, manual.categoryBtn)}>{MainCategories[category]['title']}</Button>
                                                                 );
                                                             })} */}
-                                                            <Grid item xs={12}>
+                                                            <Grid item sm={11}>
                                                                 <Field
                                                                     component={CheckboxWithLabel}
                                                                     name="needs_tag"
@@ -394,10 +400,10 @@ const ResourceFormDesktop = (props) => {
                                                             </Grid>
                                                         </div>
                                                     </Grid>
-                                                    <Grid item xs={12}>
+                                                    <Grid item sm={12}>
                                                         <FormikField label="Other Tags (Seperate each by semicolon)"
-                                                                     placeholder="Separate Each Tag by Semicolon"
-                                                                     name="other_tags" />
+                                                            placeholder="Separate Each Tag by Semicolon"
+                                                            name="other_tags" />
                                                     </Grid>
                                                 </Grid>
                                             </div>
@@ -405,8 +411,8 @@ const ResourceFormDesktop = (props) => {
                                             <div className={manual.section} style={{ marginTop: '30px' }}>
                                                 <div>
                                                     By adding a resource, you agree to the <a style={{ textAlign: "center", color: "#4284C8" }}
-                                                                                              href="https://policylibrary.columbia.edu/acceptable-usage-information-resources-policy" target="_blank">
-                                                    Columbia Resources Policy</a>.
+                                                        href="https://policylibrary.columbia.edu/acceptable-usage-information-resources-policy" target="_blank">
+                                                        Columbia Resources Policy</a>.
                                                 </div>
                                                 <Field
                                                     component={CheckboxWithLabel}
@@ -432,12 +438,13 @@ const ResourceFormDesktop = (props) => {
                                     )
                                 }}
                             </Formik>
+                        </Grid>
+                    </Grid>
                     <div style={{ marginBottom: "50px" }} />
                     {/* </div> */}
                 </Container>
             </div>
         </div >
-        </Template>
     );
 };
